requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
python-dotenv>=1.0.1
httpx>=0.27.0
jinja2>=3.1.4
pydantic-settings>=2.6.0
psycopg2-binary>=2.9.10
Pillow>=11.0.0
